ingest_api/requirements.txt
requests==2.26.0
google-cloud-bigquery==2.28.1